format: Reorganizando diretorio de testes
