--- backend/tests/unit/.env.example
@@ -1,14 +1,0 @@
-# In order to run the tests, you must setup a .env file in the tests folder
-
-# Postgres config
-POSTGRES_HOST="localhost"
-POSTGRES_PORT="7777"
-POSTGRES_PASSWORD="ubuntu"
-POSTGRES_USER="ubuntu"
-POSTGRES_DB="orb"
-POSTGRES_SLL_MODE="disable"
-
-# Migration configuration 
-MIGRATE="true"
-# Test loadEnv
-TEST="test"
