+++ backend/tests/unit/repositories/database_test.go
@@ -1,4 +1,4 @@
-package unit
+package repositories
 
 import (
 	"github.com/stretchr/testify/assert"
