+++ backend/tests/unit/repositories/user_role/user_role_test.go
@@ -1,4 +1,4 @@
-package unit
+package userrole
 
 import (
 	"github.com/stretchr/testify/assert"
